♻️ X81Y8X blueprints: decompose recipe hint resolution
