+++ .agentplane/tasks/202605290307-X81Y8X/pr/github-body.md
@@ -1,0 +1,33 @@
+Task: `202605290307-X81Y8X`
+Title: Blueprint resolve decomposition
+Canonical task record: `.agentplane/tasks/202605290307-X81Y8X/README.md`
+
+## Summary
+
+Blueprint resolve decomposition
+
+Decompose packages/agentplane/src/blueprints/resolve.ts by extracting recipe hint validation helpers while preserving blueprint resolution and explanation behavior.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/blueprints/resolve.ts by extracting recipe hint validation helpers while preserving blueprint resolution and explanation behavior.
+- Out of scope: unrelated refactors not required for "Blueprint resolve decomposition".
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T03:07:30.419Z
+- Branch: task/202605290307-X81Y8X/blueprint-resolve-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
